Add host-owned SourceCode and Diff rendering boundaries
BB had two independent code-rendering call sites — the file preview's
`@pierre/diffs` File view and the diff card's FileDiff view — and plugins
had a third: import `@pierre/diffs` directly and reconstruct the patch
normalization and code-theme wiring by hand (plugins/github did exactly
that, MutationObservers on `data-bb-code-theme-*` included).

Introduce one host boundary per capability. `SourceCodeHost` and
`DiffHost` resolve an exclusive plugin replacement through the same
resolver/PluginReplacementSlot path the sidebar thread list and file
openers use, and otherwise render BB's own renderer. Both BB renderers
sit behind `lazy()`, so a replacement that never delegates never
downloads them and `experimental_Original` costs nothing until rendered.

The renderers now own the code theme and light/dark selection themselves,
which makes the `themeType` prop threaded from ThreadTimelineSurface down
to TimelineFileDiffBlock dead; it is removed along with the opaque
`diffViewOptions` record, replaced by the semantic `DiffPresentation`
(view / overflow / showLineNumbers) that a plugin replacement can also
receive.

Public surface (all experimental, entries follow in docs/api_to_audit.md):
`experimental_SourceCode`, `experimental_Diff`,
`app.slots.experimental_sourceCodeRenderer`, and
`app.slots.experimental_diffRenderer`.
