+++ apps/app/src/components/secondary-panel/ThreadSecondaryPanelTabContent.panelGate.test.tsx
@@ -68,7 +68,11 @@ describe("GitDiffTabContent panel gating", () => {
           target={TARGET}
           isDiffPanelActive
           isPanelOpen={isPanelOpen}
-          gitDiffViewOptions={{}}
+          gitDiffPresentation={{
+          view: "unified",
+          overflow: "scroll",
+          showLineNumbers: true,
+        }}
         />
       </Wrapper>
     );
