+++ packages/plugin-sdk/src/app-contract.ts
@@ -183,6 +183,108 @@ export interface PluginFileOpenerProps {
   experimental_Original: ComponentType;
 }
 
+// ---------------------------------------------------------------------------
+// Host-owned code rendering (SourceCode / Diff) — the public components and
+// the props their replacements receive.
+// ---------------------------------------------------------------------------
+
+/** How a code line longer than the viewport is presented. */
+export type CodeOverflowMode = "scroll" | "wrap";
+
+/** How a diff presents its two sides. */
+export type DiffViewMode = "unified" | "split";
+
+/** A 1-based, inclusive line range. */
+export interface SourceCodeLineRange {
+  start: number;
+  end: number;
+}
+
+/**
+ * Props of the host-owned `experimental_SourceCode` component — BB's source
+ * viewer. The host owns syntax highlighting, gutters, wrapping, line-selection
+ * presentation, and the live BB code theme; the caller owns loading the text
+ * and any surrounding chrome.
+ */
+export interface SourceCodeProps {
+  /** The complete source text to render. */
+  content: string;
+  /** File path or name. Drives language detection and the a11y label. */
+  path: string;
+  /** Long-line presentation. Defaults to `"scroll"`. */
+  overflow?: CodeOverflowMode;
+  /**
+   * Lines to highlight and scroll into view (1-based, inclusive). Defaults to
+   * `null` — nothing highlighted.
+   */
+  highlightedLines?: SourceCodeLineRange | null;
+  /** Applied to the renderer's root element. */
+  className?: string;
+}
+
+/**
+ * Props of the host-owned `experimental_Diff` component — BB's diff viewer.
+ * The host owns patch normalization (a patch without a `diff --git` header is
+ * completed from `path`), syntax highlighting, unified/split presentation,
+ * gutters, line-selection presentation, and the live BB code theme. Content
+ * that cannot be parsed as a patch degrades to plain monospace text.
+ */
+export interface DiffProps {
+  /** Unified patch text for exactly ONE file. */
+  patch: string;
+  /**
+   * The file the patch applies to. Used to complete a patch that arrives
+   * without a `diff --git` header (GitHub's REST patches, single `@@` hunks)
+   * and for language detection.
+   */
+  path: string;
+  /** Side-by-side or inline. Defaults to `"unified"`. */
+  view?: DiffViewMode;
+  /** Long-line presentation. Defaults to `"scroll"`. */
+  overflow?: CodeOverflowMode;
+  /** Whether the gutter shows line numbers. Defaults to `true`. */
+  showLineNumbers?: boolean;
+  /** Applied to the renderer's root element. */
+  className?: string;
+}
+
+/**
+ * Props passed to an `experimental_sourceCodeRenderer` component. Every value
+ * is already resolved — the replacement never re-applies a host default.
+ */
+export interface PluginSourceCodeRendererProps {
+  content: string;
+  path: string;
+  overflow: CodeOverflowMode;
+  highlightedLines: SourceCodeLineRange | null;
+  /**
+   * BB's source renderer, bound to this request. Render it to delegate
+   * conditionally without re-entering plugin replacement resolution.
+   *
+   * @experimental Audit before relying on this as a stable contract.
+   */
+  experimental_Original: ComponentType;
+}
+
+/**
+ * Props passed to an `experimental_diffRenderer` component. `patch` is always
+ * a complete single-file unified patch, whatever shape the caller supplied.
+ */
+export interface PluginDiffRendererProps {
+  patch: string;
+  path: string;
+  view: DiffViewMode;
+  overflow: CodeOverflowMode;
+  showLineNumbers: boolean;
+  /**
+   * BB's diff renderer, bound to this request. Render it to delegate
+   * conditionally without re-entering plugin replacement resolution.
+   *
+   * @experimental Audit before relying on this as a stable contract.
+   */
+  experimental_Original: ComponentType;
+}
+
 /**
  * Message context passed to a `messageDirective` component — the assistant
  * (or nested agent) message that contained the directive.
@@ -731,6 +833,43 @@ export interface PluginFileOpenerRegistration {
 }
 
 /**
+ * Replace BB's source-code renderer everywhere it renders supplied source
+ * text — the native file preview and every plugin that calls
+ * `experimental_SourceCode`. Like `experimental_threadList` this slot is
+ * **exclusive**: one renderer at a time. Registering activates it while the
+ * plugin is enabled; if several are registered the first in deterministic slot
+ * order wins. A missing, disabled, or crashing replacement falls back to BB's
+ * renderer, and a replacement can render `experimental_Original` to delegate
+ * per call (behind its own setting, by language, by size — whatever it needs).
+ */
+export interface PluginSourceCodeRendererRegistration {
+  /** Unique within the plugin; letters, digits, `-`, `_`. */
+  id: string;
+  /** Label shown in capability details. */
+  title: string;
+  /** Optional one-line description shown with the provider choice. */
+  description?: string;
+  component: ComponentType<PluginSourceCodeRendererProps>;
+}
+
+/**
+ * Replace BB's diff renderer everywhere it renders supplied diff content — the
+ * timeline file diffs, the environment diff panel's text bodies, and every
+ * plugin that calls `experimental_Diff`. Exclusive, with the same activation,
+ * fallback, and `experimental_Original` delegation rules as
+ * {@link PluginSourceCodeRendererRegistration}.
+ */
+export interface PluginDiffRendererRegistration {
+  /** Unique within the plugin; letters, digits, `-`, `_`. */
+  id: string;
+  /** Label shown in capability details. */
+  title: string;
+  /** Optional one-line description shown with the provider choice. */
+  description?: string;
+  component: ComponentType<PluginDiffRendererProps>;
+}
+
+/**
  * Register a leaf message directive rendered inside assistant (and nested
  * agent) message Markdown. `id` is the directive name: `inline-vis` matches
  * `::inline-vis{file="demo.html"}`.
@@ -877,6 +1016,22 @@ export interface PluginAppSlots {
     registration: PluginThreadHeaderActionRegistration,
   ): void;
   fileOpener(registration: PluginFileOpenerRegistration): void;
+  /**
+   * Replace BB's source-code renderer (see
+   * {@link PluginSourceCodeRendererRegistration}). Experimental: see
+   * docs/api_to_audit.md.
+   */
+  experimental_sourceCodeRenderer(
+    registration: PluginSourceCodeRendererRegistration,
+  ): void;
+  /**
+   * Replace BB's diff renderer (see
+   * {@link PluginDiffRendererRegistration}). Experimental: see
+   * docs/api_to_audit.md.
+   */
+  experimental_diffRenderer(
+    registration: PluginDiffRendererRegistration,
+  ): void;
   messageDirective(registration: PluginMessageDirectiveRegistration): void;
   messageAction(registration: PluginMessageActionRegistration): void;
   /**
@@ -1526,5 +1681,20 @@ export interface PluginSdkApp {
    * docs/api_to_audit.md for what to audit before the prefix drops.
    */
   experimental_NewThreadComposer: ComponentType<NewThreadComposerProps>;
+  /**
+   * The host-owned source viewer (see {@link SourceCodeProps}). Renders
+   * supplied source text with BB's syntax highlighting, gutters, and live code
+   * theme, and honours an active `experimental_sourceCodeRenderer`
+   * replacement. Experimental: see docs/api_to_audit.md.
+   */
+  experimental_SourceCode: ComponentType<SourceCodeProps>;
+  /**
+   * The host-owned diff viewer (see {@link DiffProps}). Renders supplied patch
+   * content with BB's normalization, syntax highlighting, unified/split
+   * presentation, and live code theme, and honours an active
+   * `experimental_diffRenderer` replacement. Experimental: see
+   * docs/api_to_audit.md.
+   */
+  experimental_Diff: ComponentType<DiffProps>;
   useComposerView(): ComposerView;
 }
