+++ packages/plugin-sdk/src/app.ts
@@ -50,6 +50,9 @@ export const ThreadChat = runtime.ThreadChat;
 export const Markdown = runtime.Markdown;
 export const experimental_NewThreadComposer =
   runtime.experimental_NewThreadComposer;
+// Host-owned code rendering (experimental — see docs/api_to_audit.md).
+export const experimental_SourceCode = runtime.experimental_SourceCode;
+export const experimental_Diff = runtime.experimental_Diff;
 export const useRpc = runtime.useRpc;
 export const useRealtime = runtime.useRealtime;
 export const useRealtimeConnectionState = runtime.useRealtimeConnectionState;
